requirements.txt
# Core dependencies
python-dotenv
requests
ffmpeg-python
faster-whisper
huggingface_hub[hf_xet]